+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Logarithmically accelerating spinner for spaCy commands.
+* New `--debug` argument.
 
 ### Changed
+* Improved error handling.
 * Introduced a half-second input delay for "Any other text?" prompts.
 * Allowed early exit from integer prompts.
 * Wrapped words in single quotes for MFWs analysis.